+++ src/app/components/SearchResults/SearchResultsPage.jsx
@@ -5,17 +5,25 @@ import { bindActionCreators } from 'redux';
 import { withRouter } from 'react-router';
 import * as DS from '@nypl/design-system-react-components';
 import FeatureFlags from 'dgx-feature-flags';
+import { ButtonTypes } from '@nypl/design-system-react-components/lib/components/01-atoms/Button/ButtonTypes';
 import * as searchActions from '../../actions/SearchActions';
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
 import { initialSearchQuery, searchQueryPropTypes } from '../../stores/InitialState';
-import { deepEqual, isEmpty, checkFeatureFlagActivated } from '../../util/Util';
+import {
+  deepEqual, isEmpty, getNumberOfPages, checkFeatureFlagActivated,
+} from '../../util/Util';
 import TotalWorks from '../SearchForm/TotalWorks';
 
 import featureFlagConfig from '../../../../featureFlagConfig';
 import config from '../../../../appConfig';
 import { searchFields } from '../../constants/fields';
 import SearchHeader from '../SearchForm/SearchHeader';
-import SearchResults from './SearchResults';
+import { getQueryString } from '../../search/query';
+import { sortMap, numbersPerPage } from '../../constants/sorts';
+import { breakpoints } from '../../constants/breakpoints';
+import Filters from './Filters';
+import ResultsList from './ResultsList';
+import SearchPagination from './SearchPagination';
 
 export const isValidSearchQuery = query => !!query && !isEmpty(query) && !!query.queries && !isEmpty(query.queries);
 
@@ -49,6 +57,11 @@ export const loadSearch = (props, context) => {
     }
   }
 };
+// redirect to url with query params
+export const submit = (query, router) => {
+  const path = `/search?${getQueryString(query)}`;
+  router.push(path);
+};
 
 /**
  * Container class providing the Redux action creators
@@ -62,15 +75,20 @@ class SearchResultsPage extends React.Component {
   constructor(props) {
     super(props);
     const { dispatch } = props;
-    this.state = { ...props, isFeatureFlagsActivated: {} };
+    this.state = {
+      ...props, isMobile: false, isModalOpen: false, isFeatureFlagsActivated: {},
+    };
 
     this.boundActions = bindActionCreators(searchActions, dispatch);
+    this.toggleFilterMenu = this.toggleFilterMenu.bind(this);
   }
 
   componentDidMount() {
     loadSearch(this.props, this.context);
 
     FeatureFlags.store.listen(this.onFeatureFlagsChange.bind(this));
+    window.addEventListener('resize', this.onWindowResize.bind(this));
+    this.onWindowResize();
 
     checkFeatureFlagActivated(
       featureFlagConfig.featureFlagList, this.state.isFeatureFlagsActivated,
@@ -94,6 +112,16 @@ class SearchResultsPage extends React.Component {
     this.setState({ featureFlagsStore: FeatureFlags.store.getState() });
   }
 
+  onWindowResize() {
+    if (window.innerWidth < breakpoints.large) {
+      this.setState({ isMobile: true });
+      this.setState({ isModalOpen: false });
+    } else {
+      this.setState({ isMobile: false });
+      this.setState({ isModalOpen: false });
+    }
+  }
+
   getDisplayItemsHeading() {
     const queriesToShow = this.props.searchQuery && this.props.searchQuery.showQueries
       .filter(query => searchFields.includes(query.field));
@@ -104,12 +132,82 @@ class SearchResultsPage extends React.Component {
     return queries.join('');
   }
 
+  toggleFilterMenu() {
+    this.setState(prevState => ({ isModalOpen: !prevState.isModalOpen }));
+  }
+
   render() {
-    const { searchQuery, searchResults, eReaderUrl } = this.props;
-    const { router, history } = this.context;
+    const { searchQuery, eReaderUrl } = this.props;
+    const { router } = this.context;
+    const searchResults = this.props.searchResults && this.props.searchResults.data && this.props.searchResults.data.data;
+    const numberOfWorks = searchResults && searchResults.totalWorks;
+    const works = searchResults && searchResults.works;
+
+    let filterCount = 0;
+    if (searchQuery.filters) {
+      filterCount = searchQuery.filters.length;
+
+      // "Show All" doesn't appear in Filters array, but counts as a filter when checked, and doesn't count when unchecked;
+      if (searchQuery.filters.find(filter => filter.field === 'show_all')) {
+        filterCount -= 1;
+      } else {
+        filterCount += 1;
+      }
+    }
+
+    const totalPages = getNumberOfPages(numberOfWorks, searchQuery.per_page);
+    const firstElement = (Number(searchQuery.per_page || 10) * Number(searchQuery.page || 0)) + 1;
+    let lastElement = Number(searchQuery.per_page || 10) * (Number(searchQuery.page || 0) + 1) || 10;
+    if (searchQuery.page >= totalPages - 1 && lastElement > numberOfWorks) {
+      lastElement = numberOfWorks;
+    }
+
+    const itemCount = (
+      <DS.Heading
+        level={2}
+        id="page-title-heading"
+        blockName="page-title"
+      >
+        {numberOfWorks > 0
+          ? `Viewing ${firstElement.toLocaleString()} - ${lastElement.toLocaleString()} of ${numberOfWorks.toLocaleString()} items`
+          : 'Viewing 0 items'}
+      </DS.Heading>
+    );
+    const mobileItemCount = (
+      <DS.Heading
+        level={2}
+        id="page-title-heading"
+        blockName="page-title"
+      >
+        <>
+          {numberOfWorks > 0 ? `${numberOfWorks.toLocaleString()} items` : '0 items'}
+        </>
+      </DS.Heading>
+    );
+
+    const onChangePerPage = (e) => {
+      const newPage = 0;
+      const newPerPage = e.target.value;
+      if (newPerPage !== searchQuery.per_page) {
+        const newQuery = Object.assign({}, searchQuery, { page: newPage, per_page: newPerPage, total: numberOfWorks || 0 });
+        // this.props.userQuery(newQuery);
+        submit(newQuery, router);
+      }
+    };
+
+    // click and navigate with different sort
+    const onChangeSort = (e) => {
+      if (e.target.value !== Object.keys(sortMap).find(key => sortMap[key] === searchQuery.sort)) {
+        const newQuery = Object.assign({}, searchQuery, { sort: sortMap[e.target.value], page: 0 });
+        // userQuery(newQuery);
+        submit(newQuery, router);
+      }
+    };
     return (
       <DS.Container>
-        <main id="mainContent">
+        <main
+          id="mainContent"
+        >
           <Breadcrumbs
             router={router}
             location={this.props.location}
@@ -133,16 +231,96 @@ class SearchResultsPage extends React.Component {
                 text={`Search Results for ${this.getDisplayItemsHeading()}`}
               />
             </div>
-            <SearchResults
-              searchQuery={searchQuery}
-              results={searchResults.data}
-              eReaderUrl={eReaderUrl}
-              {...this.boundActions}
-              history={history}
-              router={router}
-            />
+            <div className="search-navigation">
+              {this.state.isMobile ? mobileItemCount : itemCount}
+              {this.state.isMobile && (
+                <div className="filter-refine">
+                  {filterCount !== 0 && (
+                  <span className="filter-count">
+                    {filterCount}
+                    {' '}
+                    {filterCount === 1 ? 'filter' : 'filters'}
+                  </span>
+                  )}
+                  <DS.Button
+                    id="filter-button"
+                    buttonType={ButtonTypes.Link}
+                    callback={this.toggleFilterMenu}
+                  >
+                Refine
+                  </DS.Button>
+                </div>
+              )}
+              {!this.state.isMobile && (
+              <div className="search-dropdowns">
+                <DS.Dropdown
+                  dropdownId="items-per-page-select"
+                  isRequired={false}
+                  labelPosition="left"
+                  labelText="Items Per Page"
+                  labelId="nav-items-per-page"
+                  selectedOption={searchQuery.per_page ? searchQuery.per_page : undefined}
+                  dropdownOptions={numbersPerPage.map(number => number.toString())}
+                  onSelectChange={onChangePerPage}
+                  onSelectBlur={onChangePerPage}
+                />
+                <DS.Dropdown
+                  dropdownId="sort-by-select"
+                  isRequired={false}
+                  labelPosition="left"
+                  labelText="Sort By"
+                  labelId="nav-sort-by"
+                  selectedOption={searchQuery.sort
+                    ? Object.keys(sortMap).find(key => deepEqual(sortMap[key], searchQuery.sort)) : undefined}
+                  dropdownOptions={Object.keys(sortMap).map(sortOption => sortOption)}
+                  onSelectChange={onChangeSort}
+                  onSelectBlur={onChangeSort}
+                />
+              </div>
+              )}
+            </div>
+            <div className="grid-row sfr-results-container">
+              {!this.state.isMobile && (
+              <div className="nypl-results-column">
+                <Filters
+                  toggleMenu={this.toggleFilterMenu}
+                  data={searchResults}
+                  searchQuery={searchQuery}
+                  router={router}
+                  onChangeSort={onChangeSort}
+                  onChangePerPage={onChangePerPage}
+                />
+              </div>
+              )}
+              <div className="nypl-results-main">
+                <ResultsList
+                  results={works}
+                  eReaderUrl={eReaderUrl}
+                />
+              </div>
+            </div>
+            <div className="grid-row">
+              <SearchPagination
+                totalItems={numberOfWorks}
+                searchQuery={searchQuery}
+                router={router}
+              />
+            </div>
           </div>
         </main>
+        {this.state.isMobile && this.state.isModalOpen && (
+        <DS.Modal>
+          <Filters
+            toggleMenu={this.toggleFilterMenu}
+            isMobile={this.state.isMobile}
+            data={searchResults}
+            onChangeSort={onChangeSort}
+            onChangePerPage={onChangePerPage}
+            searchQuery={searchQuery}
+            router={router}
+          />
+        </DS.Modal>
+        )}
       </DS.Container>
     );
   }
@@ -151,7 +329,6 @@ class SearchResultsPage extends React.Component {
 SearchResultsPage.propTypes = {
   searchResults: PropTypes.objectOf(PropTypes.any),
   searchQuery: searchQueryPropTypes,
-  workResult: PropTypes.objectOf(PropTypes.any),
   dispatch: PropTypes.func,
   eReaderUrl: PropTypes.string,
   location: PropTypes.objectOf(PropTypes.any),
@@ -160,7 +337,6 @@ SearchResultsPage.propTypes = {
 SearchResultsPage.defaultProps = {
   searchResults: {},
   searchQuery: initialSearchQuery,
-  workResult: {},
   dispatch: () => { },
   eReaderUrl: '',
   location: {},